// src/app/views/users/models/user.model.ts
export interface User {
    id: number;
    user_cmnd_id: string;
    user_cmnd_ngay_cap: string;
    user_cmnd_noi_cap: string;
    user_created_time: Date;
    user_dia_chi_cu_ngu_1: string;
    user_dia_chi_cu_ngu_2: string;
    user_email: string;
    user_giao_ho: string;
    user_giao_phan: string;
    user_giao_xu: string;
    user_ho_ten: string;
    user_last_login_time: string;
    user_last_updated_time: Date;
    user_ngay_sinh: string;
    user_noi_sinh: string;
    user_so_dien_thoai_1: string;
    user_so_dien_thoai_2: string;
    user_status: string;
    user_ten_thanh: string;
    username: string;
}